fix(tinygo): split os.OpenRoot usage behind !tinygo build tags
TinyGo 0.39.0 does not implement os.OpenRoot (Go 1.25 std addition).
Split the containment code into build-tagged files:

- internal/lint/rootfs.go: //go:build !tinygo (uses os.OpenRoot)
- internal/lint/rootfs_tinygo.go: //go:build tinygo (falls back to os.DirFS)
- pkg/mdsmith/workspace_fs.go: //go:build !tinygo (OSWorkspace.FS with os.OpenRoot)
- pkg/mdsmith/workspace_fs_tinygo.go: //go:build tinygo (OSWorkspace.FS with os.DirFS)

The wasm sandbox has no real filesystem symlinks, so RESOLVE_BENEATH
containment is unnecessary there. Also add unit tests covering the
os.OpenRoot error-fallback paths (openRootErrFS and workspace fallback)
to bring patch coverage to target.
