+++ internal/lint/rootfs.go
@@ -1,3 +1,5 @@
+//go:build !tinygo
+
 package lint
 
 import (
